requirements.txt
streamlit>=1.20
google-generativeai>=0.5.0
ddgs>=3.9.0
python-dotenv>=1.0.0
requests>=2.28.0

